correction de l'ensemble

+++ backend/controlers/book.js
@@ -1,4 +1,5 @@
 const Book = require('../models/Book');
+const fs = require('fs');
 
 // ajouter un élément
 exports.createBook = (req, res, next) => {
@@ -10,7 +11,6 @@ exports.createBook = (req, res, next) => {
         userId: req.auth.userId,
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
-
     book.save()
     .then(() => res.status(201).json({ message: 'Objet enregistré !'}))
     .catch(error => res.status(400).json({ error }));
@@ -34,25 +34,37 @@ exports.updateBook = (req, res, next) =>{
                 .catch(error => res.status(401).json({ error }));
             }
         })
+        .catch(error => res.status(400).json({ error }));
 };
 
 // supprimer un éléments
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id})
-        .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-        .catch(error => res.status(400).json({ error }));
+    Book.findOne({_id: req.params.id})
+    .then(book =>{
+        if(book.userId != req.auth.userId){
+            res.status(401).json({message: 'NON-Authorizé !'})
+        } else {
+            const filename = book.imageUrl.split('/images/')[1];
+            fs.unlink(`image/${filename}`, () =>{
+                Book.deleteOne({_id: req.params.id})
+                .then(() => res.status(200).json({message: 'Objet supprimé !'}))
+                .catch(error => res.status(401).json({error}));
+            });
+        };
+    })
+    .catch(error => res.status(500).json({error}));
 };
 
 // lire les information de lélément séléctionner
 exports.selectBook = (req, res, next) =>{
     Book.findOne({ _id: req.params.id})
-        .then(books => res.status(200).json(books))
+        .then(book => res.status(200).json(book))
         .catch(error => res.status(404).json({ error }));
 };
 
 // lire tout les éléments
 exports.readBook = (req, res, next) =>{
     Book.find()
-        .then(books => res.status(200).json(books))
+        .then(book => res.status(200).json(book))
         .catch(error => res.status(400).json({ error }));
 };
